chore: 구독 현황 불러오기 훅으로 변경
