+++ src/api/userData.ts
@@ -26,5 +26,6 @@ export const fetchUserData = async () => {
   }
 
   const data = await response.json();
+
   return data;
 };
